Expose global game state to the memory
This patch adds a global game state that will be easily accessible from the outside by scanning for the game state signature and then using the known field offsets.
This will notably be used for autospltting and maye other tools if needed.
The way this is done does mean we are using unsafe code but it was the best way i thought of doing this.
Obviously some things can be improved about it, maybe adding some other fields or using properties to avoid duplicate code and stuff, this works for now though and I didn't want to break the Unity part of things that i'm still pretty unfamiliar with.