GameState: Handle pointer length errors and support 64 bits pointers
This patch adds support for 64 bits (long pointers), The GameState will
detect the used pointer type on the system and use either 32 bits or 64
bits pointers. This fixes the x64 Windows build of the game.
Additionally, it will now handle errors if the pointer length is not
supported to allow the game to continue running.